tune up average_pooling_layer::back_propagation

+++ tiny_cnn/config.h
@@ -74,8 +74,8 @@ namespace tiny_cnn {
  * calculation data type
  * you can change it to float, or user defined class (fixed point,etc)
  **/
-typedef float float_t;
-//typedef double float_t;
+ //typedef float float_t;
+typedef double float_t;
 
 /**
  * size of layer, model, data etc.
